+++ hutool-core/src/main/java/cn/hutool/core/util/RuntimeUtil.java
@@ -164,4 +164,14 @@ public class RuntimeUtil {
 			process.destroy();
 		}
 	}
+	
+	/**
+	 * 增加一个JVM关闭后的钩子，用于在JVM关闭时执行某些操作
+	 * 
+	 * @param hook 钩子
+	 * @since 4.0.5
+	 */
+	public static void addShutdownHook(Runnable hook) {
+		Runtime.getRuntime().addShutdownHook((hook instanceof Thread) ? (Thread)hook : new Thread(hook));
+	}
 }