+++ hutool-http/src/test/java/cn/hutool/http/test/HttpUtilTest.java
@@ -17,7 +17,7 @@ public class HttpUtilTest {
 	@Test
 	@Ignore
 	public void getTest() {
-		String result1 = HttpUtil.get("http://photo.qzone.qq.com/fcgi-bin/fcg_list_album?uin=88888&outstyle=2", "GBK");
+		String result1 = HttpUtil.get("http://photo.qzone.qq.com/fcgi-bin/fcg_list_album?uin=88888&outstyle=2", CharsetUtil.CHARSET_GBK);
 		Console.log(result1);
 	}